feat: Created separate celery tasks per recipient per event
